refactor: Order -> Product 의존성 수정 (#66)
